strlen 영향을 받는 함수 변경

+++ ft_strjoin.c
@@ -15,8 +15,8 @@
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*new_str;
-	int		s1_len;
-	int		s2_len;
+	size_t	s1_len;
+	size_t	s2_len;
 
 	if (!(s1) && !(s2))
 		return (NULL);
